Split listeners to main and renderer
